feat: extract connected logic from KonnectorTile ✨
