+++ src/styles/connectedTile.styl
